COM-2960 move empty state display on proper cmpk
