+++ src/components/CoursesSection/index.tsx
@@ -1,10 +1,8 @@
-import { View, Text, FlatList, StyleProp, ViewStyle, TouchableOpacity, Image } from 'react-native';
-import { useNavigation } from '@react-navigation/native';
+import { View, Text, FlatList, StyleProp, ViewStyle } from 'react-native';
 import styles from './styles';
 import { CourseType, ProgramType, SubProgramType } from '../../types/CourseTypes';
 import { NextSlotsStepType } from '../../types/StepTypes';
 import { formatWordToPlural } from '../../core/helpers/utils';
-import NiPrimaryButton from '../../components/form/PrimaryButton';
 
 export const COURSE_SECTION = 'FORMATION';
 export const EVENT_SECTION = 'ÉVÉNEMENT';
@@ -14,7 +12,6 @@ type CoursesSectionProps = {
   title: string,
   type?: string,
   countStyle: StyleProp<ViewStyle>
-  showCatalogButton?: boolean,
   renderItem: (item) => JSX.Element,
 }
 
@@ -23,11 +20,8 @@ const CoursesSection = ({
   title,
   type = COURSE_SECTION,
   countStyle,
-  showCatalogButton = false,
   renderItem,
 }: CoursesSectionProps) => {
-  const navigation = useNavigation();
-
   const renderSeparator = () => <View style={styles.separator} />;
 
   return (
@@ -39,14 +33,6 @@ const CoursesSection = ({
       <FlatList horizontal data={items} keyExtractor={item => `${title}${item._id}`}
         contentContainerStyle={styles.container} renderItem={({ item }) => renderItem(item)}
         showsHorizontalScrollIndicator={false} ItemSeparatorComponent={renderSeparator} />
-      {showCatalogButton &&
-        <TouchableOpacity style={styles.courseContainer} onPress={() => navigation.navigate('Catalog')}>
-          <Text style={styles.text}>Vous n’avez pas de formation en cours...</Text>
-          <NiPrimaryButton caption="Chercher une formation" onPress={() => navigation.navigate('Catalog')} />
-          <Image source={require('../../../assets/images/aux_detective.png')} style={styles.image}
-            resizeMode='contain' />
-        </TouchableOpacity>
-      }
     </>
   );
 };
