+++ src/components/CoursesSection/styles.ts
@@ -1,7 +1,6 @@
 import { StyleSheet } from 'react-native';
-import { BORDER_RADIUS, BORDER_WIDTH, MAIN_MARGIN_LEFT, MARGIN, PADDING } from '../../styles/metrics';
+import { BORDER_RADIUS, MAIN_MARGIN_LEFT, MARGIN, PADDING } from '../../styles/metrics';
 import { FIRA_SANS_BOLD } from '../../styles/fonts';
-import { TRANSPARENT_GREY, WHITE } from '../../styles/colors';
 
 const styles = StyleSheet.create({
   container: {
@@ -23,28 +22,6 @@ const styles = StyleSheet.create({
   title: {
     ...FIRA_SANS_BOLD.LG,
     marginLeft: MAIN_MARGIN_LEFT,
-  },
-  text: {
-    ...FIRA_SANS_BOLD.MD,
-    textAlign: 'center',
-  },
-  image: {
-    position: 'absolute',
-    right: -160,
-    bottom: 8,
-    height: 144,
-  },
-  courseContainer: {
-    borderRadius: BORDER_RADIUS.SM,
-    borderWidth: BORDER_WIDTH,
-    borderColor: TRANSPARENT_GREY,
-    overflow: 'hidden',
-    backgroundColor: WHITE,
-    marginHorizontal: MAIN_MARGIN_LEFT,
-    alignItems: 'center',
-    justifyContent: 'space-between',
-    height: 176,
-    padding: PADDING.LG,
   },
 });
